deletion of own sql statements and implementing them with sqlalchemy

+++ application/threads/views.py
@@ -15,7 +15,8 @@ from application.threads.forms import EditThreadTitleForm, EditThreadDescription
 @app.route("/threads", methods=["GET"])
 def threads_index():
     categoryform = CategoryForm(None, False)
-    return render_template("threads/list.html", threads = Thread.connect_threads_and_categories(None), categoryform = categoryform)
+
+    return render_template("threads/list.html", threads = Thread.query.all(), categoryform = categoryform)
 
 @app.route("/threads/category", methods=["POST"])
 def threads_with_category():
@@ -26,7 +27,13 @@ def threads_with_category():
 
 @app.route("/threads/<category_id>", methods=["GET"])
 def open_threads_with_category(category_id):
-    return render_template("threads/list.html", threads = Thread.connect_threads_and_categories(category_id), categoryform = CategoryForm(None, False))
+    
+    help = Category.query.get(category_id).threads
+    threads = []
+    for thread_category in help: 
+        threads.append(thread_category.thread)
+    
+    return render_template("threads/list.html", threads = threads, categoryform = CategoryForm(None, False))
 
 @app.route("/threads/read/<thread_id>", methods=["POST"])
 @login_required
@@ -35,9 +42,9 @@ def threads_comment(thread_id):
     
     if not form.validate():
         thread = Thread.query.get(thread_id)
-        comments = Comment.connect_users_and_comments(thread_id)
-        user = User.query.get(thread.user_id)
-        return render_template("threads/showthread.html", form = form, comments = comments, thread = thread, user = user)
+        comments = thread.comments
+        
+        return render_template("threads/showthread.html", form = form, comments = comments, thread = thread, user = thread.user)
 
     comment = Comment(form.comment.data)
     comment.thread_id = thread_id
@@ -50,9 +57,8 @@ def threads_comment(thread_id):
 @app.route("/threads/read/<thread_id>", methods=["GET"])
 def threads_open(thread_id):
     thread = Thread.query.get(thread_id)
-    comments = Comment.connect_users_and_comments(thread_id)
 
-    
+    comments = thread.comments
     
     return render_template("threads/showthread.html", form = CommentForm(), comments = comments, thread = thread, user = thread.user)
 
@@ -62,14 +68,18 @@ def threads_openedit(thread_id):
     thread = Thread.query.get(thread_id)
     if(thread.user_id != current_user.id):
         return threads_openmythreads()
-    return render_template("threads/edit.html", thread = thread, titleform = EditThreadTitleForm(), descform = EditThreadDescriptionForm())
+    return render_template("threads/edit.html", thread = thread, titleform = EditThreadTitleForm(), descform = EditThreadDescriptionForm(), categoryform = CategoryForm(None, True))
 
 @app.route("/threads/edit/<thread_id>/conf_title", methods=["POST"])
 @login_required
 def threads_confirm_title_edit(thread_id):
     thread = Thread.query.get(thread_id)
+
+    if(thread.user_id != current_user.id):
+        return threads_openmythreads()
+
     titleform = EditThreadTitleForm(request.form)
-    print(titleform.validate())
+
     if not titleform.validate():
         
         return render_template("threads/edit.html", thread = thread, titleform = titleform, descform = EditThreadDescriptionForm())
@@ -86,10 +96,12 @@ def threads_confirm_title_edit(thread_id):
 @login_required
 def threads_confirm_description_edit(thread_id):
     thread = Thread.query.get(thread_id)
+    if(thread.user_id != current_user.id):
+        return threads_openmythreads()
     descform = EditThreadDescriptionForm(request.form)
     if not descform.validate():
         
-        return render_template("threads/edit.html", thread = thread, titleform = EditThreadTitleForm, descform = descform)
+        return render_template("threads/edit.html", thread = thread, titleform = EditThreadTitleForm(), descform = descform)
 
 
     
@@ -99,11 +111,50 @@ def threads_confirm_description_edit(thread_id):
 
     return threads_open(thread_id)
 
+@app.route("/threads/edit/<thread_id>/addcategory", methods=["POST"])
+@login_required
+def threads_add_new_category(thread_id):
+    categoryform = CategoryForm(request.form, True)
+    thread = Thread.query.get(thread_id)
+    
+    if categoryform.categories.data == 'other':
+
+        if not categoryform.validate():
+            return render_template("threads/edit.html", threadform = EditThreadTitleForm(), descform = EditThreadDescriptionForm(),categoryform = categoryform)
+
+        
+        category = Category.query.filter_by(name=categoryform.othercategory.data).first()
+
+        if category:
+            categoryform.othercategory.errors.append('Category is already on the list, please select it from there')
+            return render_template("threads/edit.html", threadform = EditThreadTitleForm(), descform = EditThreadDescriptionForm(),categoryform = categoryform)
+
+        newcategory = Category(categoryform.othercategory.data)
+        
+        db.session().add(newcategory)
+        db.session().commit()
+        thread_category = Thread_Category(thread.id, newcategory.id)
+        db.session().add(thread_category)
+        db.session().commit()
+
+        return redirect(url_for("threads_index"))
+
+    
+    
+    
+    
+    thread_category = Thread_Category(thread.id, int(categoryform.categories.data))
+    db.session().add(thread_category)
+    db.session().commit()
+
+    return redirect(url_for("threads_index"))
 
 @app.route("/threads/edit/<thread_id>/delete", methods=["POST"])
 @login_required
 def threads_delete(thread_id):
-    
+    thread = Thread.query.get(thread_id)
+    if(thread.user_id != current_user.id):
+        return threads_openmythreads()
     Comment.query.filter_by(thread_id=thread_id).delete()
     Thread_Category.query.filter_by(thread_id=thread_id).delete()
     Thread.query.filter_by(id=thread_id).delete()
@@ -113,15 +164,19 @@ def threads_delete(thread_id):
 @app.route("/threads/deletecomment/<comment_id>", methods=["POST"])
 @login_required
 def comment_delete(comment_id):
+    comment = Comment.query.get(comment_id)
+    threadid = comment.thread_id
+
     Comment.query.filter_by(id=comment_id).delete()
     db.session().commit()
-    return redirect(url_for('threads_index'))
+
+    return threads_open(threadid)
 
 
 
-@app.route("/threads/new")
+@app.route("/threads/new", methods=["GET"])
 @login_required
-def threads_form():
+def threads_new():
     categoryform = CategoryForm(None, True)
     return render_template("threads/new.html", threadform = NewThreadForm(), categoryform = categoryform)
 
@@ -133,7 +188,6 @@ def threads_create():
     threadform = NewThreadForm(request.form)
     categoryform = CategoryForm(request.form, True)
 
-    
 
     if not threadform.validate():
         return render_template("threads/new.html", threadform = threadform, categoryform = categoryform)
@@ -147,6 +201,11 @@ def threads_create():
             return render_template("threads/new.html", threadform = threadform, categoryform = categoryform)
 
         
+        category = Category.query.filter_by(name=categoryform.othercategory.data).first()
+
+        if category:
+            categoryform.othercategory.errors.append('Category is already on the list, please select it from there')
+            return render_template("threads/new.html", threadform = threadform, categoryform = categoryform)
 
         newcategory = Category(categoryform.othercategory.data)
